backend/requirements.txt
flask
easyocr
pillow
numpy
flask-cors
